Added color for inserting some vortices and plotting
We still need to figure out where to put the different functions, as methods in the classes or as stand alone scripts. Important with good documentation.

I have added code to insert single and double vortices in an angle field and made code to plot the result using pyplot. Next up is implementing complex numbers, the evolution algorithm in accordance with the documentation and loops etc in three dimensions.