np.fft replaced by sp.fft
Also, I benchmarked the hydrodynamic model against the legacy code after
fixing an error in the calc_force_density function. The results show
that an initial dipole in a nx=41 system will annihilate in about 600
time units with rho0=2**-10, gamma_S=2**-10, meaning that
the hydrodynamic model is implemented correctly.

A lot of stuff could still be done with improving the efficiency of the
code with priority on base system.